This commit adds some features and cleans up some code to make it more reusable 1. Refactors logic into FastCollection which now takes an arbitrary model and view 2. Added public/private to things in a reasonable way 3. Adding the ability to have variable row heights 4. Making row height a parameter you can set 5. Removing a VStack which was at the base of ContentView that didn't do anything 6. Added a buffer parameter which allows you to load more views above and below the content window and means you're less likely to see white space when scrolling quickly
